hdat: Disable TM on Power9 DD 2.1
Update pa_features_p9[] to disable TM (Transactional Memory). On DD 2.1
TM is not usable by Linux without other workarounds, so skiboot must
disable it.

The presence of TM is communicated by setting bit 7 of byte 22 in the
pa-features array. As no other bits are set in that byte, we currently
have a value of 0x80. To disable TM we set bit 7 to 0, leaving a value
of 0x0.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>