gard: create: Allow creating arbitrary GARD records
Add a new sub-command that allows us to create GARD records for
arbitrary chip units. There isn't a whole lot of constraints on this and
that limits how useful it can be, but it does allow a user to GARD out
individual DIMMs, chips or cores from the BMC (or host) if needed.

There are a few caveats though:

1) Not everything can, or should, have a GARD record applied it to.
2) There is no validation that the unit actually exists. Doing that
   sort of validation requires something that understands the FAPI
   targeting information (I think) and adding support for it here
   would require some knowledge from the system XML file.
3) There's no way to get a list of paths in the system.
4) Although we can create a GARD record at runtime it won't be applied
   until the next IPL.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>