npu2: Remove side effects in assert() calls.
This likely looks intentional and will do the correct thing.

The correct behaviour does rely on the fact that skiboot has a custom
assert() which will always generate code. 'Regular' assert() can be made
to generate no code. If we were to do that with this code, the results
would be, interesting!

This patch will also keep Coverity happy which is nice.

Fixes: CID 145242, 172044 and 172045
Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>