p8-i2c: Don't write the watermark register at init
On P9 the I2C master is shared with the OCC. Currently the watermark
values are set once at init time which is bad for two reasons:

a) We don't take the OCC master lock before setting it. Which
   may cause issues if the OCC is currently using the master.
b) The OCC might change the watermark levels and we need to reset
   them.

Change this so that we set the watermark value when a new transaction
is started rather than at init time.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>