npu2: Refactor BAR setting code
This refactors the BAR setting code to make it clearer and handle a
larger range of BAR addresses. This is needed as we are about to move
the GPU to a physical address that is currently not supported by this
code.

This change derives group and chip sections of the BAR from the base
address rather than the chip_id now.  mem sel is also derived from the
base address, rather than assuming 0.

No functional change.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Reviewed-by: Balbir Singh <bsingharora@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>
(cherry picked from commit 3c0408ded5a14e110c8a418be305ac20714cb32d)
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>