xive: Implement "single escalation" feature
This adds a new VP flag to control the new DD2.0
"single escalation" feature.

This feature allows us to have a single escalation
interrupt per VP instead of one per queue.

It works by hijacking queue 7 (which is this no longer
usable when that is enabled) and exploiting two new
hardware bits that will:

  - Make the normal queues (0..6) escalate unconditionally
    thus ignoring the ESe bits.

  - Route the above escalations to queue 7

  - Have queue 7 silently escalate without notification

Thus the escalation of queue 7 becomes the one escalation
interrupt for all the other queues.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>