npu2: hw-procedures: Change phy_rx_clock_sel values
The clock selection bits we set here are inputs to a state machine.

  DL clock select (bits 30-31)
    0b00: lane 0 clock
    0b01: lane 7 clock
    0b10: grid clock
    0b11: invalid/noop

To recover from a potential glitch, we need to ensure that the value we
set forces a state change. Our current sequence is to set 0x3 followed
by 0x1. With the above now known, that is actually a noop followed by
selection of lane 7. Depending on lane reversal, that selection is not a
state change for some bricks.

The way to force a state change in all cases is to switch to the grid
clock, and then back to a lane.

Signed-off-by: Reza Arbab <arbab@linux.vnet.ibm.com>
Acked-by: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>
(cherry picked from commit 878c718aed200cc2b6b7c6bca3a6e2fa2351ec95)
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>