fast-reboot: add sreset_all_others error handling
Pass back failures from sreset_all_others, also change return codes to
OPAL_ form in sreset_all_prepare to match.

Errors will revert to the IPL path, so it's not critical to completely
clean up everything if that would complicate things. Detecting the
error and failing is the important thing.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>