libflash/mbox-flash: Allow mbox-flash to tell the driver msg timeouts
Currently when mbox-flash decides that a message times out the driver
has no way of knowing to drop the message and will continue waiting for
a response indefinitely preventing more messages from ever being sent.

This is a problem if the BMC crashes or has some other issue where it
won't ever respond to our outstanding message.

This patch provides a method for mbox-flash to tell the driver how long
it should wait before it no longer needs to care about the response.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>