libflash/mbox-flash: fallback to requesting lower MBOX versions from BMC
Some BMC mbox implementations seem to sometimes mysteriously fail when trying
to negotiate v3 when they only support v2. To work around this, we
can fall back to requesting lower mbox protocol versions until we find
one that works.

In theory, this should already "just work", but we have a counter example,
which this patch fixes.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>