pci: Make fast reboot creset PHBs in parallel
In this patch, we added a creset to the fast reboot path:

  commit a1eba9d29d17ff44fca1293b071023a9fbf4b938
  Author: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
  fast-reboot: creset PHBs on fast reboot

Unfortunately, this does the creset in serial rather than in
parallel.

This patch changes the creset on fast reboot to operate in
parallel. It simplifies the code by abstracting pci_init_slots() so
that it can be used on boot and fast reboot to perform all the PCI
retraining.

This improves fast reboot PCI training on a dual chip P9 from ~12
seconds down to ~2 seconds. Similar improvements for P8 are achieved
as well.

This means the time to download the BOOTKERNEL is the bottleneck for
fast reboot.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Reviewed-by: Russell Currey <ruscur@russell.cc>
Acked-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>