README: document output files
A grumpy BMC developer was wondering which file to flash to the machines
they were testing on. They longed for the old days where the answer was
'skiboot.lid'. Alas, those days are long past.

This updates the readme with a list of the current output formats and
where you might want to use them.

Signed-off-by: Joel Stanley <joel@jms.id.au>
[stewart: fix P8/P9 thing]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>